requirements.txt
prettytable==0.7.2
trakt==2.10.0
pyyaml==5.1.1
docopt==0.6.2
yaspin==0.14.3
termcolor==1.1.0
python-crontab==2.3.8
plexapi==3.1.0
arrow==0.14.3
